hermes_service/requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
Flask==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
priority==2.0.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyTelegramBotAPI==3.6.6
python-dotenv==1.0.1
Quart==0.19.6
redis==4.5.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
taskgroup==0.0.0a4
tomli==2.0.1
twilio==9.1.0
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
